+++ src/AuthCallback.tsx
@@ -1,0 +1,16 @@
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+const AuthCallback = () => {
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    navigate('/', { replace: true });
+  }, []);
+
+  return (
+    <span>processing authentication...</span>
+  );
+};
+
+export default AuthCallback;
